🔧fix: Change model data type in GRPO trainer initialization
- Updated the model data type from "bfloat16" to "float16" in `grpo_trainer.py` to improve compatibility and performance.
- This change aligns with recent updates to the GRPO configuration, enhancing the overall training process.

+++ rlhf/grpo/grpo_trainer.py
@@ -243,7 +243,7 @@ class UnslothGRPOTrainer:
             self.model, self.tokenizer = FastLanguageModel.from_pretrained(
                 model_name=model_name,
                 max_seq_length=getattr(self.config, 'max_prompt_length', 2048),
-                dtype=torch.bfloat16,
+                dtype=torch.float16,
                 load_in_4bit=True,
                 device_map="balanced"
             )
